backend/requirements.txt
# RAG Ingestion Pipeline Dependencies

cohere>=5.0.0
qdrant-client>=1.7.0
langchain>=0.1.0
langchain-community>=0.0.10
python-dotenv>=1.0.0
tiktoken>=0.5.0

# RAG Backend API Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
google-genai>=0.3.0
pydantic>=2.6.0
grok>=1.0.0
